# PERF: Use fused types for map_infer_mask