--- a/pandas/_libs/dtypes.pxd
+++ b/pandas/_libs/dtypes.pxd
@@ -34,3 +34,8 @@ ctypedef fused numeric_t:
 ctypedef fused numeric_object_t:
     numeric_t
     object
+
+ctypedef fused uint8_int64_object_t:
+    uint8_t
+    int64_t
+    object